took out a separate component LeaderboardTableHead
