+++ src/entities/leaderboard/lib/sort-table-by-column.ts
@@ -5,14 +5,12 @@ export const sortTableByColumn = (
   order: TOrder,
   orderBy: keyof TDataRowLeaderboardTable,
 ): TDataRowLeaderboardTable[] => {
-  const arrThis = array;
-
-  arrThis.sort((a, b) => {
+  array.sort((a, b) => {
     if (order === 'desc') {
       return a[orderBy] < b[orderBy] ? 1 : -1;
     }
     return a[orderBy] > b[orderBy] ? 1 : -1;
   });
 
-  return arrThis;
+  return array;
 };
