+++ src/entities/leaderboard/ui/leaderboard-table-head/index.ts
@@ -1,0 +1,1 @@
+export { LeaderboardTableHead } from './leaderboard-table-head';
