+++ src/entities/leaderboard/ui/leaderboard-table/leaderboard-table.tsx
@@ -5,89 +5,26 @@ import {
   TableBody,
   TableCell,
   TableContainer,
-  TableHead,
   TablePagination,
   TableRow,
-  TableSortLabel,
   Typography,
 } from '@mui/material';
-import {
-  TDataRowLeaderboardTable,
-  TDataRowsLeaderboardTable,
-  TEnhancedTableProps,
-  THeadCell,
-  TOrder,
-} from 'entities/leaderboard/types';
+import { TDataRowLeaderboardTable, TOrder } from 'entities/leaderboard/types';
 import { sortTableByColumn } from 'entities/leaderboard/lib';
-
-const headCells: readonly THeadCell[] = [
-  {
-    id: 'displayName',
-    numeric: false,
-    disablePadding: false,
-    align: 'left',
-    label: 'Никнейм',
-  },
-  {
-    id: 'allPlays',
-    numeric: true,
-    disablePadding: false,
-    align: 'right',
-    label: 'Кол-во игр',
-  },
-  {
-    id: 'winsCount',
-    numeric: true,
-    disablePadding: false,
-    align: 'right',
-    label: 'Кол-во побед',
-  },
-  {
-    id: 'winsPercent',
-    numeric: true,
-    disablePadding: false,
-    align: 'right',
-    label: 'Процент побед',
-  },
-];
-
-const EnhancedTableHead = ({ order, orderBy, onRequestSort }: TEnhancedTableProps) => {
-  const createSortHandler =
-    (property: keyof TDataRowLeaderboardTable) => (event: MouseEvent<unknown>) => {
-      onRequestSort(event, property);
-    };
-
-  return (
-    <TableHead>
-      <TableRow>
-        {headCells.map((headCell) => (
-          <TableCell
-            key={headCell.id}
-            align={headCell.align}
-            padding={headCell.disablePadding ? 'none' : 'normal'}
-            sortDirection={orderBy === headCell.id ? order : false}
-          >
-            <TableSortLabel
-              active={orderBy === headCell.id}
-              direction={orderBy === headCell.id ? order : 'asc'}
-              onClick={createSortHandler(headCell.id)}
-            >
-              {headCell.label}
-            </TableSortLabel>
-          </TableCell>
-        ))}
-      </TableRow>
-    </TableHead>
-  );
-};
+import { LeaderboardTableHead } from 'entities/leaderboard/ui/leaderboard-table-head';
 
 const INITIAL_PAGE_INDEX = 0;
+const INITIAL_ROW_PER_PAGE = 5;
 
-export const LeaderboardTable = ({ dataTable }: TDataRowsLeaderboardTable) => {
+type Props = {
+  dataTable: TDataRowLeaderboardTable[];
+};
+
+export const LeaderboardTable = ({ dataTable }: Props) => {
   const [order, setOrder] = useState<TOrder>('desc');
   const [orderBy, setOrderBy] = useState<keyof TDataRowLeaderboardTable>('winsCount');
-  const [page, setPage] = useState(0);
-  const [rowsPerPage, setRowsPerPage] = useState(5);
+  const [page, setPage] = useState(INITIAL_PAGE_INDEX);
+  const [rowsPerPage, setRowsPerPage] = useState(INITIAL_ROW_PER_PAGE);
 
   const handleRequestSort = (
     event: MouseEvent<unknown>,
@@ -120,14 +57,13 @@ export const LeaderboardTable = ({ dataTable }: TDataRowsLeaderboardTable) => {
         id="tableTitle"
         component="div"
         align="center"
-        pt={2}
-        pb={2}
+        py={2}
       >
         Таблица победителей
       </Typography>
       <TableContainer sx={{ maxHeight: 440 }}>
         <Table sx={{ minWidth: 750 }} aria-labelledby="tableTitle" size="medium">
-          <EnhancedTableHead order={order} orderBy={orderBy} onRequestSort={handleRequestSort} />
+          <LeaderboardTableHead order={order} orderBy={orderBy} onRequestSort={handleRequestSort} />
           <TableBody>
             {sortTableByColumn(dataTable, order, orderBy)
               .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
